tiler: avoid lock-ups due to unmapped DMM entries
Due to the address remapping done for 2d buffers, when a 2d buffer is
allocated by userspace, it is possible that only partial pages map to
valid physical pages of memory.  But the MMU can only protect accesses
to invalid addresses on the granualarity of full pages.  Which results
in the inconvenient situation where access to a perfectly valid address
(from the point of view of a userspace process's memory map) can lock
up the processor.

To protect against this, a single dummy page is allocated at boot time
and mapped in to the entire DMM space.  When actual buffers are alloc'd,
the mapping is replaced with the actual pages allocated for the buffer.
And when the actual TILER buffer is freed, the mappings are set back to
the dummy page.

As a result, there is no longer the need for a clear() function in DMM.
Instead the clear operation is simply remapping DMM entries back to the
dummy page.

As an added advantage, now pages are properly unmapped from DMM when
TILER buffers are freed, avoiding the potential to corrupt pages that
where previously allocated for TILER buffers.