omap:dss:hdmi: tame hdmi driver
Main change is to avoid automatically picking timings.  Now the driver
defaults to a safe 640x480, and leaves it to the upper layer (drm driver)
to pick the desired resolution based on what the user selects and the
drm infrastucture's parsing of the EDID.

Note: some of this may need to be checked w/ S3D team, and other use cases
where android is using the HDMI display.. it is still the correct approach
for the HDMI driver to not automatically pick timings, but possibly there
needs to be something in userspace to do this if x11 is not running.  See
the modetest test app coming from libdrm tree.. something like this could
be used to set the modeline from userspace.