Support of multiple sync objects of SGXSubmitTransferKM
This is a fixed and optimized version of SGXSubmitTransferKM
patch provided by IMG. This change allows for multiple source
and destination sync objects to be supported allowing for a
more efficient surface management. The fixes and optimizations
were in the array use. More optimizations are required to
reduce the overhead of the iterations traversing all sync
objects.

Change-Id: I415228ba3eef33552d2a6275428f0e399d9ac37e
Reviewed-on: http://dlegerr.ext.ti.com/gerrit/omap/203
Tested-by: Ian Elliott <ian.elliott@intelligraphics.com>
Reviewed-by: Ian Elliott <ian.elliott@intelligraphics.com>
Reviewed-by: Jeff McGee <j-mcgee@ti.com>